+++ clients/admin-ui/src/features/common/table/cells/types.d.ts
@@ -1,0 +1,5 @@
+export interface ColumnState {
+  isExpanded?: boolean;
+  isWrapped?: boolean;
+  version?: number;
+}