+++ clients/fidesui/src/index.ts
@@ -17,6 +17,7 @@ export type {
 export type {
   ButtonProps as AntButtonProps,
   CollapseProps as AntCollapseProps,
+  DropdownProps as AntDropdownProps,
   FlexProps as AntFlexProps,
   FormInstance as AntFormInstance,
   InputProps as AntInputProps,
@@ -57,7 +58,6 @@ export {
   Menu as AntMenu,
   message as AntMessage,
   Modal as AntModal,
-  Pagination as AntPagination,
   Radio as AntRadio,
   Row as AntRow,
   Skeleton as AntSkeleton,
@@ -82,6 +82,7 @@ export type { UploadChangeParam as AntUploadChangeParam } from "antd/lib/upload"
 // Higher-order components
 export {
   CustomDateRangePicker as AntDateRangePicker,
+  CustomPagination as AntPagination,
   CustomSelect as AntSelect,
   CustomTable as AntTable,
   CustomTag as AntTag,
@@ -89,6 +90,9 @@ export {
   CustomTypography as AntTypography,
 } from "./hoc";
 export type { CustomTagProps as AntTagProps } from "./hoc/CustomTag";
+
+// Constants
+export { PAGE_SIZES } from "./hoc/CustomPagination";
 
 // Export the destructured Typography components individually
 export const AntText = CustomTypography.Text;